pythonWebSocketServer/requirements.txt
simplejson
websockets
kafka-python